prepare for merge with renames
