+++ maggy/core/experiment_driver/ablation_driver.py
@@ -21,7 +21,7 @@ from maggy.earlystop import NoStoppingRule
 from maggy.ablation.ablationstudy import AblationStudy
 from maggy.ablation.ablator.loco import LOCO
 from maggy.ablation.ablator import AbstractAblator
-from maggy.core.experiment_driver.OptimizationDriver import OptimizationDriver
+from maggy.core.experiment_driver.optimization_driver import OptimizationDriver
 
 
 class AblationDriver(OptimizationDriver):
